+++ src/test/java/com/a6raywa1cher/coursejournalbackend/integration/models/SubmissionInfo.java
@@ -3,5 +3,6 @@ package com.a6raywa1cher.coursejournalbackend.integration.models;
 import java.time.ZonedDateTime;
 import java.util.List;
 
-public record SubmissionInfo(long task, List<Long> satisfiedCriteria, ZonedDateTime submittedAt, int additionalScore) {
+public record SubmissionInfo(long task, List<Long> satisfiedCriteria, ZonedDateTime submittedAt,
+                             double additionalScore) {
 }
